requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pipenv==2023.11.17
pippip==0.0.10
platformdirs==4.1.0
protobuf==4.25.2
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ultralytics==8.1.6
urllib3==2.1.0
validators==0.22.0
virtualenv==20.25.0
watchdog==3.0.0
zipp==3.17.0
